⚙️reorganized file structure and content
moved branch specific files into its own folders
split client events up

**made it look less like a mess and more like angular.js**
