+++ clientFiles/public/index.html
@@ -12,9 +12,13 @@
       crossorigin="anonymous"
     ></script>
     <script src='/socket.js'></script><!--Needs to be 1. Position-->
+
     <script src='/borderFunctions.js'></script><!--Files with Fuction Defs-->
+    <script src='/movement.js'></script>
     <script src='/main.js'></script>
-    <script src='/events.js'></script><!--Event file needs to be last-->
+
+    <script src='/MovementCollisionEvents.js'></script>
+    <script src='/Events.js'></script><!--Event file needs to be last--><!--Event files needs to be last-->
     <title>AmongHTL</title>
   </head>
   <body style="margin: 0px">
