+++ router.js
@@ -1,24 +1,40 @@
 module.exports = function (app) {
 	
+//html
+
     app.get('/', function(req, res) {
         res.sendFile('/public/index.html', {root: 'clientFiles' });
     });
     
+//scripts
+
     app.get('/main.js', function(req, res) {
         res.sendFile('/src/main.js', {root: 'clientFiles' });
-    });
-
-    app.get('/borderFunctions.js', function(req, res) {
-        res.sendFile('/src/borderFunctions.js', {root: 'clientFiles' });
     });
 
     app.get('/socket.js', function(req, res) {
         res.sendFile('/src/socket.js', {root: 'clientFiles' });
     });
 
-    app.get('/events.js', function(req, res) {
-        res.sendFile('/src/events.js', {root: 'clientFiles' });
+    app.get('/Events.js', function(req, res) {
+        res.sendFile('/src/Events.js', {root: 'clientFiles' });
     });
+
+    //Movement Collision Scripts
+    app.get('/borderFunctions.js', function(req, res) {
+        res.sendFile('/src/Movement_Collision/borderFunctions.js', {root: 'clientFiles' });
+    });
+
+    app.get('/movement.js', function(req, res) {
+        res.sendFile('/src/Movement_Collision/movement.js', {root: 'clientFiles' });
+    });
+
+    app.get('/MovementCollisionEvents.js', function(req, res) {
+        res.sendFile('/src/Movement_Collision/MovementCollisionEvents.js', {root: 'clientFiles' });
+    });
+
+//css
+
     
 //images
